refactor: TE7-T119 Active Record Callbacks
refactor a code in module class.
